[FIX] git not being able to commit without setting "user.name" and "user.email"
Apparently, GIT_AUTHOR_NAME is not enough. In some cases, git expects
the GIT_COMMITTER_NAME env variable.

We could also set EMAIL env variable, but it's not necessary.
It fallbacks to $USER@$HOSTNAME (ie: odoo@b412358129138).
And we really don't care, we just want to be able to commit.