fix: multi-arch images
Multi architecture images were being built in parallel jobs, and each job pushed
the manifest to the registry.

As the manifest for each architecture share the same tag, there was a race
condition between the jobs, and the last one to push the manifest would win.

This caused our images to actually be single-arch.

The solution is to push the manifests by digest (without tagging), and then have
a separate job create a manifest list merging the different arch images and
pushing it with the proper tag.