🔥 entrypoint: remove create extension unaccent
This entrypoint is not necessary.

It was used to create the unaccent extension on the database, in case it wasn't
already created. However, Odoo already takes care of this when initializing new
databases. The only case not covered by Odoo is when switching ON this setting
on an existing database, but that case should not be covered by an entrypoint.